+++ propertybasedtests/rnkata_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 var cases = []struct {
-	Arabic int
+	Arabic uint16
 	Roman  string
 }{
 	{Arabic: 1, Roman: "I"},
